refactor: move dpi method
